Move Feature flat file.
