--- src/Libraries/Domains/Vendors/Confie.WesternGeneral/FlatFile/Feature.cs
@@ -1,77 +1,0 @@
-﻿using System;
-using FileHelpers;
-using FileHelpers.Events;
-
-namespace Confie.WesternGeneral.FlatFile
-{
-    [FixedLengthRecord]
-    public class Feature : INotifyRead
-    {
-        [FieldFixedLength(50)]
-        [FieldTrim(TrimMode.Right)]
-        public string ClaimId { get; set; }
-
-        [FieldFixedLength(50)]
-        [FieldTrim(TrimMode.Right)]
-        public string FeatureId { get; set; }
-
-        [FieldFixedLength(18)]
-        [FieldConverter(ConverterKind.Date, "yyyy/MM/dd hh:mmtt")]
-        public DateTime OpenDate { get; set; }
-
-        [FieldFixedLength(18)]
-        [FieldNullValue(typeof(DateTime), "0001/01/01 12:00AM")]
-        [FieldConverter(ConverterKind.Date, "yyyy/MM/dd hh:mmtt")]
-        public DateTime CloseDate { get; set; }
-
-        [FieldFixedLength(30)]
-        [FieldTrim(TrimMode.Right)]
-        public string FeatureStatus { get; set; }
-
-        [FieldFixedLength(30)]
-        [FieldTrim(TrimMode.Right)]
-        public string CoverageCode { get; set; }
-
-        [FieldFixedLength(30)]
-        [FieldTrim(TrimMode.Right)]
-        public string CoverageSubCode { get; set; }
-
-        [FieldFixedLength(1)]
-        [FieldConverter(ConverterKind.Boolean)]
-        public bool ClosedWithoutPayment { get; set; }
-
-        [FieldFixedLength(17)]
-        [FieldConverter(ConverterKind.Decimal)]
-        public decimal Payments { get; set; }
-
-        [FieldFixedLength(17)]
-        [FieldConverter(ConverterKind.Decimal)]
-        public decimal Recoveries { get; set; }
-
-        [FieldFixedLength(17)]
-        [FieldConverter(ConverterKind.Decimal)]
-        public decimal Expenses { get; set; }
-
-        [FieldFixedLength(17)]
-        [FieldConverter(ConverterKind.Decimal)]
-        public decimal ReservesAtBeginning { get; set; }
-
-        [FieldFixedLength(17)]
-        [FieldConverter(ConverterKind.Decimal)]
-        public decimal ReservesAtEnd { get; set; }
-
-        [FieldFixedLength(17)]
-        [FieldConverter(ConverterKind.Decimal)]
-        public decimal TotalIncurredLoss { get; set; }
-
-        public void BeforeRead(BeforeReadEventArgs e)
-        {
-            e.SkipThisRecord = string.IsNullOrWhiteSpace(e.RecordLine) || e.RecordLine.Equals("\u001a");
-        }
-
-        public void AfterRead(AfterReadEventArgs e)
-        {
-            //Do nothing.
-        }
-    }
-}
